requirements.txt
Flask==1.1.1
flask-restplus==0.13.0
Werkzeug==0.15.4
opencv-python==3.4.4.19
pandas==0.24.0
Pillow==5.3.0
scikit-learn==0.21.3
joblib==0.14.0
sklearn==0.0
virtualenv==16.0.0
